refactor(serverless): clean code
