+++ packages/serverless/.env.example
@@ -1,6 +1,6 @@
 LAGON_TOKEN=
-LAGON_API_URL=http://localhost:3000/api
 LAGON_ROOT_DOMAIN=lagon.app
+LAGON_WORKERS=1
 SENTRY_DSN=
 SENTRY_ENVIRONMENT=development
