+++ packages/serverless/src/constants.ts
@@ -1,4 +1,1 @@
-import { cpus } from 'node:os';
-
 export const IS_DEV = process.env.NODE_ENV !== 'production';
-export const CPUS = cpus().length;
